fixed torch.inverse for float16

+++ rome/compute_u.py
@@ -49,8 +49,8 @@ def get_inv_cov(
             precision=mom2_dtype,
         )
         inv_mom2_cache[key] = torch.inverse(
-            stat.mom2.moment().to("cuda")
-        ).float()  # Cast back to float32
+            stat.mom2.moment().float().to("cuda")
+        ).half()  # Cast back to float32
 
     return inv_mom2_cache[key]
